Add prune command to delete old job items
Bug: https://phabricator.wikimedia.org/T183338
