+++ bin/ia-upload
@@ -5,6 +5,7 @@ require __DIR__ . '/../vendor/autoload.php';
 
 use Wikisource\IaUpload\Command\JobsCommand;
 use Symfony\Component\Console\Application;
+use Wikisource\IaUpload\Command\PruneCommand;
 
 $configFile = __DIR__ . '/../config.ini';
 $config = parse_ini_file( $configFile );
@@ -17,4 +18,5 @@ setlocale( LC_ALL, 'en_US.UTF-8' );
 
 $application = new Application();
 $application->add( new JobsCommand( $config ) );
+$application->add( new PruneCommand( $config ) );
 $application->run();
